+++ rscx/src/format_wrapper.rs
@@ -1,18 +1,19 @@
 use std::fmt::{Display, self};
 
-use crate::renderable::Renderable;
+use crate::render::Render;
 
-pub struct FormatWrapper<T: Renderable> {
+pub struct FormatWrapper<T> {
     inner: T,
 }
 
-impl<T: Renderable> FormatWrapper<T> {
+impl<T> FormatWrapper<T> {
     pub fn new(inner: T) -> Self {
         Self { inner }
     }
 }
 
-impl<T: Renderable> Display for FormatWrapper<T> {
+impl<T: Render> Display for FormatWrapper<T> {
+    #[inline(always)]
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.inner.render(f)
     }
